hearing-rehab-app/requirements.txt
# Flask 相關
Flask==2.3.3
python-dotenv==1.0.0

# Gemini TTS 相關
google-generativeai==0.3.2

# Qwen3-TTS 相關 (可選)
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.37.0
soundfile>=0.12.1
librosa>=0.10.0
numpy>=1.24.0

# 其他工具
requests>=2.31.0
Pillow>=10.0.0